fix: some renaming and refactoring
