+++ tee-worker/ts-tests/common/utils/index.ts
@@ -8,6 +8,7 @@ export * from './assertion';
 export * from './common';
 export * from './context';
 export * from './crypto';
-export * from './identity';
+export * from './identity-helper';
+export * from './integration-setup';
 export * from './storage';
-export * from './test';
+export * from './vc-helper';
